added login and register routes

+++ database/migrations/2018_12_11_183946_add_authenticated_to_clients_table.php
@@ -26,7 +26,9 @@ class AddAuthenticatedToClientsTable extends Migration
     public function down()
     {
         Schema::table('clients', function (Blueprint $table) {
-            $table->dropColumn('authenticated');
+            if (Schema::hasColumn('clients', 'authenticated')) {
+                $table->dropColumn('authenticated');
+            }
         });
     }
 }
