+++ database/migrations/2018_12_11_184004_add_password_to_clients_table.php
@@ -26,7 +26,9 @@ class AddPasswordToClientsTable extends Migration
     public function down()
     {
         Schema::table('clients', function (Blueprint $table) {
-            $table->dropColumn('password');
+            if (Schema::hasColumn('clients', 'password')) {
+                $table->dropColumn('password');
+            }
         });
     }
 }
